Renderer: Introduce expectation paradigm to tests
Instead of providing the exactly output of the function, the testing code
automatically generates the actual image and compare it with the expected,
generating a diff in case of disagreement.

If the developer updates an algorithm which have a different, but desired,
expectation, the developer can replace the old expectation with the new
one easily.