Renderer: Fix the line renderer
Previously we always rendered the line by iterating the x-axis, the issue
arise in lines that have a high slope or even are vertical.

In high slopes the x-axis would be smaller, giving not enough resolution
to draw the line. This get extra bad if the line is vertical, we would have
a division by zero error.

Now we always render the line in the longest axis