Change some Ordering
I changed the Ordering::Release to Ordering::AcqRel because in my current
mental model of how ordering works, we need to pair an Acquire with a Release
to stablish happens-before relationship.

In this case we do not want the thread to decrement the counter before
executing the task, thus needing a 'Release' semantic. Also we do not want
the finished_mutex being written to 'true' before decrementing the counter
(due to an especulative write?). It might not be even possible, since it's
behind a lock and the memory model prevents reordering of atomics but since
I'm still learning, I'm better be safe.